Refactor H12 and H1X (#503)
* refactor h12 and h1x

* fix linting

* fix legacy tests

* wip migrate tests

* wip migrate tests

* fix tests

* migrate h1x tests;

* fix tests

* fix test
