+++ malariagen_data/anopheles.py
@@ -2,7 +2,6 @@ import math
 import warnings
 from abc import abstractmethod
 from bisect import bisect_left, bisect_right
-from collections import Counter
 from typing import Any, Dict, List, Mapping, Optional, Tuple, Sequence
 
 import allel  # type: ignore
@@ -33,7 +32,6 @@ from .anoph import (
     dash_params,
     diplotype_distance_params,
     gplt_params,
-    h12_params,
     hapclust_params,
     hapnet_params,
     het_params,
@@ -54,6 +52,8 @@ from .anoph.sample_metadata import AnophelesSampleMetadata, locate_cohorts
 from .anoph.snp_data import AnophelesSnpData
 from .anoph.g123 import AnophelesG123Analysis
 from .anoph.fst import AnophelesFstAnalysis
+from .anoph.h12 import AnophelesH12Analysis
+from .anoph.h1x import AnophelesH1XAnalysis
 from .mjn import median_joining_network, mjn_graph
 from .util import (
     CacheMiss,
@@ -71,7 +71,6 @@ from .util import (
     region_str,
     simple_xarray_concat,
     pandas_apply,
-    hash_columns,
 )
 
 DEFAULT_MAX_COVERAGE_VARIANCE = 0.2
@@ -99,6 +98,8 @@ DEFAULT_MAX_COVERAGE_VARIANCE = 0.2
 # work around pycharm failing to recognise that doc() is callable
 # noinspection PyCallingNonCallable
 class AnophelesDataResource(
+    AnophelesH1XAnalysis,
+    AnophelesH12Analysis,
     AnophelesG123Analysis,
     AnophelesFstAnalysis,
     AnophelesSnpFrequencyAnalysis,
@@ -181,23 +182,8 @@ class AnophelesDataResource(
 
     @property
     @abstractmethod
-    def _h12_calibration_cache_name(self):
-        raise NotImplementedError("Must override _h12_calibration_cache_name")
-
-    @property
-    @abstractmethod
-    def _h12_gwss_cache_name(self):
-        raise NotImplementedError("Must override _h12_gwss_cache_name")
-
-    @property
-    @abstractmethod
     def _xpehh_gwss_cache_name(self):
         raise NotImplementedError("Must override _xpehh_gwss_cache_name")
-
-    @property
-    @abstractmethod
-    def _h1x_gwss_cache_name(self):
-        raise NotImplementedError("Must override _h1x_gwss_cache_name")
 
     @property
     @abstractmethod
@@ -1845,742 +1831,6 @@ class AnophelesDataResource(
             return None
         else:
             return (fig1, fig2, fig3, fig4)
-
-    @check_types
-    @doc(
-        summary="Generate h12 GWSS calibration data for different window sizes.",
-        returns="""
-            A list of H12 calibration run arrays for each window size, containing
-            values and percentiles.
-        """,
-    )
-    def h12_calibration(
-        self,
-        contig: base_params.contig,
-        analysis: hap_params.analysis = DEFAULT,
-        sample_query: Optional[base_params.sample_query] = None,
-        sample_sets: Optional[base_params.sample_sets] = None,
-        cohort_size: Optional[base_params.cohort_size] = h12_params.cohort_size_default,
-        min_cohort_size: Optional[
-            base_params.min_cohort_size
-        ] = h12_params.min_cohort_size_default,
-        max_cohort_size: Optional[
-            base_params.max_cohort_size
-        ] = h12_params.max_cohort_size_default,
-        window_sizes: h12_params.window_sizes = h12_params.window_sizes_default,
-        random_seed: base_params.random_seed = 42,
-    ) -> Mapping[str, np.ndarray]:
-        # change this name if you ever change the behaviour of this function, to
-        # invalidate any previously cached data
-        name = self._h12_calibration_cache_name
-
-        params = dict(
-            contig=contig,
-            analysis=self._prep_phasing_analysis_param(analysis=analysis),
-            window_sizes=window_sizes,
-            sample_sets=self._prep_sample_sets_param(sample_sets=sample_sets),
-            # N.B., do not be tempted to convert this sample query into integer
-            # indices using _prep_sample_selection_params, because the indices
-            # are different in the haplotype data.
-            sample_query=sample_query,
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            random_seed=random_seed,
-        )
-
-        try:
-            calibration_runs = self.results_cache_get(name=name, params=params)
-
-        except CacheMiss:
-            calibration_runs = self._h12_calibration(**params)
-            self.results_cache_set(name=name, params=params, results=calibration_runs)
-
-        return calibration_runs
-
-    def _h12_calibration(
-        self,
-        contig,
-        analysis,
-        sample_query,
-        sample_sets,
-        cohort_size,
-        min_cohort_size,
-        max_cohort_size,
-        window_sizes,
-        random_seed,
-    ) -> Mapping[str, np.ndarray]:
-        # access haplotypes
-        ds_haps = self.haplotypes(
-            region=contig,
-            sample_sets=sample_sets,
-            sample_query=sample_query,
-            analysis=analysis,
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            random_seed=random_seed,
-        )
-
-        gt = allel.GenotypeDaskArray(ds_haps["call_genotype"].data)
-        with self._dask_progress(desc="Load haplotypes"):
-            ht = gt.to_haplotypes().compute()
-
-        calibration_runs: Dict[str, np.ndarray] = dict()
-        for window_size in self._progress(window_sizes, desc="Compute H12"):
-            h12 = allel.moving_statistic(ht, statistic=_garud_h12, size=window_size)
-            calibration_runs[str(window_size)] = h12
-
-        return calibration_runs
-
-    @check_types
-    @doc(
-        summary="Plot h12 GWSS calibration data for different window sizes.",
-        parameters=dict(
-            title="Plot title.",
-            show="If True, show the plot.",
-        ),
-    )
-    def plot_h12_calibration(
-        self,
-        contig: base_params.contig,
-        analysis: hap_params.analysis = DEFAULT,
-        sample_query: Optional[base_params.sample_query] = None,
-        sample_sets: Optional[base_params.sample_sets] = None,
-        cohort_size: Optional[base_params.cohort_size] = h12_params.cohort_size_default,
-        min_cohort_size: Optional[
-            base_params.min_cohort_size
-        ] = h12_params.min_cohort_size_default,
-        max_cohort_size: Optional[
-            base_params.max_cohort_size
-        ] = h12_params.max_cohort_size_default,
-        window_sizes: h12_params.window_sizes = h12_params.window_sizes_default,
-        random_seed: base_params.random_seed = 42,
-        title: Optional[str] = None,
-        show: bool = True,
-    ) -> gplt_params.figure:
-        # get H12 values
-        calibration_runs = self.h12_calibration(
-            contig=contig,
-            analysis=analysis,
-            sample_query=sample_query,
-            sample_sets=sample_sets,
-            window_sizes=window_sizes,
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            random_seed=random_seed,
-        )
-
-        # compute summaries
-        q50 = [np.median(calibration_runs[str(window)]) for window in window_sizes]
-        q25 = [
-            np.percentile(calibration_runs[str(window)], 25) for window in window_sizes
-        ]
-        q75 = [
-            np.percentile(calibration_runs[str(window)], 75) for window in window_sizes
-        ]
-        q05 = [
-            np.percentile(calibration_runs[str(window)], 5) for window in window_sizes
-        ]
-        q95 = [
-            np.percentile(calibration_runs[str(window)], 95) for window in window_sizes
-        ]
-
-        # make plot
-        if title is None:
-            title = sample_query
-        fig = bokeh.plotting.figure(
-            title=title,
-            width=700,
-            height=400,
-            x_axis_type="log",
-            x_range=bokeh.models.Range1d(window_sizes[0], window_sizes[-1]),
-        )
-        fig.patch(
-            window_sizes + window_sizes[::-1],
-            q75 + q25[::-1],
-            alpha=0.75,
-            line_width=2,
-            legend_label="25-75%",
-        )
-        fig.patch(
-            window_sizes + window_sizes[::-1],
-            q95 + q05[::-1],
-            alpha=0.5,
-            line_width=2,
-            legend_label="5-95%",
-        )
-        fig.line(
-            window_sizes, q50, line_color="black", line_width=4, legend_label="median"
-        )
-        fig.circle(window_sizes, q50, color="black", fill_color="black", size=8)
-
-        fig.xaxis.ticker = window_sizes
-        if show:  # pragma: no cover
-            bokeh.plotting.show(fig)
-            return None
-        else:
-            return fig
-
-    @check_types
-    @doc(
-        summary="Run h12 genome-wide selection scan.",
-        returns=dict(
-            x="An array containing the window centre point genomic positions.",
-            h12="An array with h12 statistic values for each window.",
-        ),
-    )
-    def h12_gwss(
-        self,
-        contig: base_params.contig,
-        window_size: h12_params.window_size,
-        analysis: hap_params.analysis = DEFAULT,
-        sample_query: Optional[base_params.sample_query] = None,
-        sample_sets: Optional[base_params.sample_sets] = None,
-        cohort_size: Optional[base_params.cohort_size] = h12_params.cohort_size_default,
-        min_cohort_size: Optional[
-            base_params.min_cohort_size
-        ] = h12_params.min_cohort_size_default,
-        max_cohort_size: Optional[
-            base_params.max_cohort_size
-        ] = h12_params.max_cohort_size_default,
-        random_seed: base_params.random_seed = 42,
-    ) -> Tuple[np.ndarray, np.ndarray]:
-        # change this name if you ever change the behaviour of this function, to
-        # invalidate any previously cached data
-        name = self._h12_gwss_cache_name
-
-        params = dict(
-            contig=contig,
-            analysis=self._prep_phasing_analysis_param(analysis=analysis),
-            window_size=window_size,
-            sample_sets=self._prep_sample_sets_param(sample_sets=sample_sets),
-            # N.B., do not be tempted to convert this sample query into integer
-            # indices using _prep_sample_selection_params, because the indices
-            # are different in the haplotype data.
-            sample_query=sample_query,
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            random_seed=random_seed,
-        )
-
-        try:
-            results = self.results_cache_get(name=name, params=params)
-
-        except CacheMiss:
-            results = self._h12_gwss(**params)
-            self.results_cache_set(name=name, params=params, results=results)
-
-        x = results["x"]
-        h12 = results["h12"]
-
-        return x, h12
-
-    def _h12_gwss(
-        self,
-        contig,
-        analysis,
-        window_size,
-        sample_sets,
-        sample_query,
-        cohort_size,
-        min_cohort_size,
-        max_cohort_size,
-        random_seed,
-    ):
-        ds_haps = self.haplotypes(
-            region=contig,
-            analysis=analysis,
-            sample_query=sample_query,
-            sample_sets=sample_sets,
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            random_seed=random_seed,
-        )
-
-        gt = allel.GenotypeDaskArray(ds_haps["call_genotype"].data)
-        with self._dask_progress(desc="Load haplotypes"):
-            ht = gt.to_haplotypes().compute()
-
-        with self._spinner(desc="Compute H12"):
-            # Compute H12.
-            h12 = allel.moving_statistic(ht, statistic=_garud_h12, size=window_size)
-
-            # Compute window midpoints.
-            pos = ds_haps["variant_position"].values
-            x = allel.moving_statistic(pos, statistic=np.mean, size=window_size)
-
-        results = dict(x=x, h12=h12)
-
-        return results
-
-    @check_types
-    @doc(
-        summary="Plot h12 GWSS data.",
-    )
-    def plot_h12_gwss_track(
-        self,
-        contig: base_params.contig,
-        window_size: h12_params.window_size,
-        analysis: hap_params.analysis = DEFAULT,
-        sample_sets: Optional[base_params.sample_sets] = None,
-        sample_query: Optional[base_params.sample_query] = None,
-        cohort_size: Optional[base_params.cohort_size] = h12_params.cohort_size_default,
-        min_cohort_size: Optional[
-            base_params.min_cohort_size
-        ] = h12_params.min_cohort_size_default,
-        max_cohort_size: Optional[
-            base_params.max_cohort_size
-        ] = h12_params.max_cohort_size_default,
-        random_seed: base_params.random_seed = 42,
-        title: Optional[gplt_params.title] = None,
-        sizing_mode: gplt_params.sizing_mode = gplt_params.sizing_mode_default,
-        width: gplt_params.width = gplt_params.width_default,
-        height: gplt_params.height = 200,
-        show: gplt_params.show = True,
-        x_range: Optional[gplt_params.x_range] = None,
-        output_backend: gplt_params.output_backend = gplt_params.output_backend_default,
-    ) -> gplt_params.figure:
-        # compute H12
-        x, h12 = self.h12_gwss(
-            contig=contig,
-            analysis=analysis,
-            window_size=window_size,
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            sample_query=sample_query,
-            sample_sets=sample_sets,
-            random_seed=random_seed,
-        )
-
-        # determine X axis range
-        x_min = x[0]
-        x_max = x[-1]
-        if x_range is None:
-            x_range = bokeh.models.Range1d(x_min, x_max, bounds="auto")
-
-        # create a figure
-        xwheel_zoom = bokeh.models.WheelZoomTool(
-            dimensions="width", maintain_focus=False
-        )
-        if title is None:
-            title = sample_query
-        fig = bokeh.plotting.figure(
-            title=title,
-            tools=[
-                "xpan",
-                "xzoom_in",
-                "xzoom_out",
-                xwheel_zoom,
-                "reset",
-                "save",
-                "crosshair",
-            ],
-            active_inspect=None,
-            active_scroll=xwheel_zoom,
-            active_drag="xpan",
-            sizing_mode=sizing_mode,
-            width=width,
-            height=height,
-            toolbar_location="above",
-            x_range=x_range,
-            y_range=(0, 1),
-            output_backend=output_backend,
-        )
-
-        # plot H12
-        fig.circle(
-            x=x,
-            y=h12,
-            size=3,
-            line_width=1,
-            line_color="black",
-            fill_color=None,
-        )
-
-        # tidy up the plot
-        fig.yaxis.axis_label = "H12"
-        fig.yaxis.ticker = [0, 1]
-        self._bokeh_style_genome_xaxis(fig, contig)
-
-        if show:  # pragma: no cover
-            bokeh.plotting.show(fig)
-            return None
-        else:
-            return fig
-
-    @check_types
-    @doc(
-        summary="Plot h12 GWSS data.",
-    )
-    def plot_h12_gwss(
-        self,
-        contig: base_params.contig,
-        window_size: h12_params.window_size,
-        analysis: hap_params.analysis = DEFAULT,
-        sample_sets: Optional[base_params.sample_sets] = None,
-        sample_query: Optional[base_params.sample_query] = None,
-        cohort_size: Optional[base_params.cohort_size] = h12_params.cohort_size_default,
-        min_cohort_size: Optional[
-            base_params.min_cohort_size
-        ] = h12_params.min_cohort_size_default,
-        max_cohort_size: Optional[
-            base_params.max_cohort_size
-        ] = h12_params.max_cohort_size_default,
-        random_seed: base_params.random_seed = 42,
-        title: Optional[gplt_params.title] = None,
-        sizing_mode: gplt_params.sizing_mode = gplt_params.sizing_mode_default,
-        width: gplt_params.width = gplt_params.width_default,
-        track_height: gplt_params.track_height = 170,
-        genes_height: gplt_params.genes_height = gplt_params.genes_height_default,
-        show: gplt_params.show = True,
-        output_backend: gplt_params.output_backend = gplt_params.output_backend_default,
-    ) -> gplt_params.figure:
-        # gwss track
-        fig1 = self.plot_h12_gwss_track(
-            contig=contig,
-            analysis=analysis,
-            window_size=window_size,
-            sample_sets=sample_sets,
-            sample_query=sample_query,
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            random_seed=random_seed,
-            title=title,
-            sizing_mode=sizing_mode,
-            width=width,
-            height=track_height,
-            show=False,
-            output_backend=output_backend,
-        )
-
-        fig1.xaxis.visible = False
-
-        # plot genes
-        fig2 = self.plot_genes(
-            region=contig,
-            sizing_mode=sizing_mode,
-            width=width,
-            height=genes_height,
-            x_range=fig1.x_range,
-            show=False,
-            output_backend=output_backend,
-        )
-
-        # combine plots into a single figure
-        fig = bokeh.layouts.gridplot(
-            [fig1, fig2],
-            ncols=1,
-            toolbar_location="above",
-            merge_tools=True,
-            sizing_mode=sizing_mode,
-        )
-
-        if show:  # pragma: no cover
-            bokeh.plotting.show(fig)
-            return None
-        else:
-            return fig
-
-    @check_types
-    @doc(
-        summary="""
-            Run a H1X genome-wide scan to detect genome regions with
-            shared selective sweeps between two cohorts.
-        """,
-        returns=dict(
-            x="An array containing the window centre point genomic positions.",
-            h1x="An array with H1X statistic values for each window.",
-        ),
-    )
-    def h1x_gwss(
-        self,
-        contig: base_params.contig,
-        window_size: h12_params.window_size,
-        cohort1_query: base_params.sample_query,
-        cohort2_query: base_params.sample_query,
-        analysis: hap_params.analysis = DEFAULT,
-        sample_sets: Optional[base_params.sample_sets] = None,
-        cohort_size: Optional[base_params.cohort_size] = h12_params.cohort_size_default,
-        min_cohort_size: Optional[
-            base_params.min_cohort_size
-        ] = h12_params.min_cohort_size_default,
-        max_cohort_size: Optional[
-            base_params.max_cohort_size
-        ] = h12_params.max_cohort_size_default,
-        random_seed: base_params.random_seed = 42,
-    ) -> Tuple[np.ndarray, np.ndarray]:
-        # change this name if you ever change the behaviour of this function, to
-        # invalidate any previously cached data
-        name = self._h1x_gwss_cache_name
-
-        params = dict(
-            contig=contig,
-            analysis=self._prep_phasing_analysis_param(analysis=analysis),
-            window_size=window_size,
-            # N.B., do not be tempted to convert these sample queries into integer
-            # indices using _prep_sample_selection_params, because the indices
-            # are different in the haplotype data.
-            cohort1_query=cohort1_query,
-            cohort2_query=cohort2_query,
-            sample_sets=self._prep_sample_sets_param(sample_sets=sample_sets),
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            random_seed=random_seed,
-        )
-
-        try:
-            results = self.results_cache_get(name=name, params=params)
-
-        except CacheMiss:
-            results = self._h1x_gwss(**params)
-            self.results_cache_set(name=name, params=params, results=results)
-
-        x = results["x"]
-        h1x = results["h1x"]
-
-        return x, h1x
-
-    def _h1x_gwss(
-        self,
-        contig,
-        analysis,
-        window_size,
-        sample_sets,
-        cohort1_query,
-        cohort2_query,
-        cohort_size,
-        min_cohort_size,
-        max_cohort_size,
-        random_seed,
-    ):
-        # access haplotype datasets for each cohort
-        ds1 = self.haplotypes(
-            region=contig,
-            analysis=analysis,
-            sample_query=cohort1_query,
-            sample_sets=sample_sets,
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            random_seed=random_seed,
-        )
-        ds2 = self.haplotypes(
-            region=contig,
-            analysis=analysis,
-            sample_query=cohort2_query,
-            sample_sets=sample_sets,
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            random_seed=random_seed,
-        )
-
-        # load data into memory
-        gt1 = allel.GenotypeDaskArray(ds1["call_genotype"].data)
-        with self._dask_progress(desc="Load haplotypes for cohort 1"):
-            ht1 = gt1.to_haplotypes().compute()
-        gt2 = allel.GenotypeDaskArray(ds2["call_genotype"].data)
-        with self._dask_progress(desc="Load haplotypes for cohort 2"):
-            ht2 = gt2.to_haplotypes().compute()
-
-        with self._spinner(desc="Compute H1X"):
-            # Run H1X scan.
-            h1x = _moving_h1x(ht1, ht2, size=window_size)
-
-            # Compute window midpoints.
-            pos = ds1["variant_position"].values
-            x = allel.moving_statistic(pos, statistic=np.mean, size=window_size)
-
-        results = dict(x=x, h1x=h1x)
-
-        return results
-
-    @check_types
-    @doc(
-        summary="""
-            Run and plot a H1X genome-wide scan to detect genome regions
-            with shared selective sweeps between two cohorts.
-        """
-    )
-    def plot_h1x_gwss_track(
-        self,
-        contig: base_params.contig,
-        window_size: h12_params.window_size,
-        cohort1_query: base_params.cohort1_query,
-        cohort2_query: base_params.cohort2_query,
-        analysis: hap_params.analysis = DEFAULT,
-        sample_sets: Optional[base_params.sample_sets] = None,
-        cohort_size: Optional[base_params.cohort_size] = h12_params.cohort_size_default,
-        min_cohort_size: Optional[
-            base_params.min_cohort_size
-        ] = h12_params.min_cohort_size_default,
-        max_cohort_size: Optional[
-            base_params.max_cohort_size
-        ] = h12_params.max_cohort_size_default,
-        random_seed: base_params.random_seed = 42,
-        title: Optional[gplt_params.title] = None,
-        sizing_mode: gplt_params.sizing_mode = gplt_params.sizing_mode_default,
-        width: gplt_params.width = gplt_params.width_default,
-        height: gplt_params.height = 200,
-        show: gplt_params.show = True,
-        x_range: Optional[gplt_params.x_range] = None,
-        output_backend: gplt_params.output_backend = gplt_params.output_backend_default,
-    ) -> gplt_params.figure:
-        # compute H1X
-        x, h1x = self.h1x_gwss(
-            contig=contig,
-            analysis=analysis,
-            window_size=window_size,
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            cohort1_query=cohort1_query,
-            cohort2_query=cohort2_query,
-            sample_sets=sample_sets,
-            random_seed=random_seed,
-        )
-
-        # determine X axis range
-        x_min = x[0]
-        x_max = x[-1]
-        if x_range is None:
-            x_range = bokeh.models.Range1d(x_min, x_max, bounds="auto")
-
-        # create a figure
-        xwheel_zoom = bokeh.models.WheelZoomTool(
-            dimensions="width", maintain_focus=False
-        )
-        if title is None:
-            title = f"Cohort 1: {cohort1_query}\nCohort 2: {cohort2_query}"
-        fig = bokeh.plotting.figure(
-            title=title,
-            tools=[
-                "xpan",
-                "xzoom_in",
-                "xzoom_out",
-                xwheel_zoom,
-                "reset",
-                "save",
-                "crosshair",
-            ],
-            active_inspect=None,
-            active_scroll=xwheel_zoom,
-            active_drag="xpan",
-            sizing_mode=sizing_mode,
-            width=width,
-            height=height,
-            toolbar_location="above",
-            x_range=x_range,
-            y_range=(0, 1),
-            output_backend=output_backend,
-        )
-
-        # plot H1X
-        fig.circle(
-            x=x,
-            y=h1x,
-            size=3,
-            line_width=1,
-            line_color="black",
-            fill_color=None,
-        )
-
-        # tidy up the plot
-        fig.yaxis.axis_label = "H1X"
-        fig.yaxis.ticker = [0, 1]
-        self._bokeh_style_genome_xaxis(fig, contig)
-
-        if show:  # pragma: no cover
-            bokeh.plotting.show(fig)
-            return None
-        else:
-            return fig
-
-    @check_types
-    @doc(
-        summary="""
-            Run and plot a H1X genome-wide scan to detect genome regions
-            with shared selective sweeps between two cohorts.
-        """
-    )
-    def plot_h1x_gwss(
-        self,
-        contig: base_params.contig,
-        window_size: h12_params.window_size,
-        cohort1_query: base_params.cohort1_query,
-        cohort2_query: base_params.cohort2_query,
-        analysis: hap_params.analysis = DEFAULT,
-        sample_sets: Optional[base_params.sample_sets] = None,
-        cohort_size: Optional[base_params.cohort_size] = h12_params.cohort_size_default,
-        min_cohort_size: Optional[
-            base_params.min_cohort_size
-        ] = h12_params.min_cohort_size_default,
-        max_cohort_size: Optional[
-            base_params.max_cohort_size
-        ] = h12_params.max_cohort_size_default,
-        random_seed: base_params.random_seed = 42,
-        title: Optional[gplt_params.title] = None,
-        sizing_mode: gplt_params.sizing_mode = gplt_params.sizing_mode_default,
-        width: gplt_params.width = gplt_params.width_default,
-        track_height: gplt_params.track_height = 190,
-        genes_height: gplt_params.genes_height = gplt_params.genes_height_default,
-        show: gplt_params.show = True,
-        output_backend: gplt_params.output_backend = gplt_params.output_backend_default,
-    ) -> gplt_params.figure:
-        # gwss track
-        fig1 = self.plot_h1x_gwss_track(
-            contig=contig,
-            analysis=analysis,
-            window_size=window_size,
-            cohort1_query=cohort1_query,
-            cohort2_query=cohort2_query,
-            sample_sets=sample_sets,
-            cohort_size=cohort_size,
-            min_cohort_size=min_cohort_size,
-            max_cohort_size=max_cohort_size,
-            random_seed=random_seed,
-            title=title,
-            sizing_mode=sizing_mode,
-            width=width,
-            height=track_height,
-            show=False,
-            output_backend=output_backend,
-        )
-
-        fig1.xaxis.visible = False
-
-        # plot genes
-        fig2 = self.plot_genes(
-            region=contig,
-            sizing_mode=sizing_mode,
-            width=width,
-            height=genes_height,
-            x_range=fig1.x_range,
-            show=False,
-            output_backend=output_backend,
-        )
-
-        # combine plots into a single figure
-        fig = bokeh.layouts.gridplot(
-            [fig1, fig2],
-            ncols=1,
-            toolbar_location="above",
-            merge_tools=True,
-            sizing_mode=sizing_mode,
-        )
-
-        if show:  # pragma: no cover
-            bokeh.plotting.show(fig)
-            return None
-        else:
-            return fig
 
     @check_types
     @doc(
@@ -4477,87 +3727,6 @@ def _unrooted_tree_layout_equal_angle(
         if leaf_colors is not None:
             leaf_color = leaf_colors.iloc[tree_node.index]
         leaf_nodes.append([x, y, tree_node.index, leaf_color])
-
-
-def _haplotype_frequencies(h):
-    """Compute haplotype frequencies, returning a dictionary that maps
-    haplotype hash values to frequencies."""
-    n = h.shape[1]
-    hashes = hash_columns(np.asarray(h))
-    counts = Counter(hashes)
-    freqs = {key: count / n for key, count in counts.items()}
-    return freqs
-
-
-def _garud_h12(ht):
-    """Compute Garud's H12."""
-
-    # compute diplotype frequencies
-    frq_counter = _haplotype_frequencies(ht)
-
-    # convert to array of sorted frequencies
-    f = np.sort(np.fromiter(frq_counter.values(), dtype=float))[::-1]
-
-    # compute H12
-    h12 = np.sum(f[:2]) ** 2 + np.sum(f[2:] ** 2)
-
-    return h12
-
-
-def _haplotype_joint_frequencies(ha, hb):
-    """Compute the joint frequency of haplotypes in two difference
-    cohorts. Returns a dictionary mapping haplotype hash values to
-    the product of frequencies in each cohort."""
-    frqa = _haplotype_frequencies(ha)
-    frqb = _haplotype_frequencies(hb)
-    keys = set(frqa.keys()) | set(frqb.keys())
-    joint_freqs = {key: frqa.get(key, 0) * frqb.get(key, 0) for key in keys}
-    return joint_freqs
-
-
-def _h1x(ha, hb):
-    """Compute H1X, the sum of joint haplotype frequencies between
-    two cohorts, which is a summary statistic useful for detecting
-    shared selective sweeps."""
-    jf = _haplotype_joint_frequencies(ha, hb)
-    return np.sum(list(jf.values()))
-
-
-def _moving_h1x(ha, hb, size, start=0, stop=None, step=None):
-    """Compute H1X in moving windows.
-
-    Parameters
-    ----------
-    ha : array_like, int, shape (n_variants, n_haplotypes)
-        Haplotype array for the first cohort.
-    hb : array_like, int, shape (n_variants, n_haplotypes)
-        Haplotype array for the second cohort.
-    size : int
-        The window size (number of variants).
-    start : int, optional
-        The index at which to start.
-    stop : int, optional
-        The index at which to stop.
-    step : int, optional
-        The number of variants between start positions of windows. If not
-        given, defaults to the window size, i.e., non-overlapping windows.
-
-    Returns
-    -------
-    h1x : ndarray, float, shape (n_windows,)
-        H1X values (sum of squares of joint haplotype frequencies).
-    """
-
-    assert ha.ndim == hb.ndim == 2
-    assert ha.shape[0] == hb.shape[0]
-
-    # construct moving windows
-    windows = allel.index_windows(ha, size, start, stop, step)
-
-    # compute statistics for each window
-    out = np.array([_h1x(ha[i:j], hb[i:j]) for i, j in windows])
-
-    return out
 
 
 @numba.njit("Tuple((int8, int64))(int8[:], int8)")
